+++ features/step_definitions/generate.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 Then("the product {string} with destination {string} contains the framework {string} with architecture {string}") do |product, destination, framework, architecture|
   framework_path = Xcode.find_framework(
     product: product,
@@ -22,4 +24,25 @@ Then("the product {string} with destination {string} contains the framework {str
   out, err, status = Open3.capture3("file", binary_path)
   assert(status.success?, err)
   refute(out.include?(architecture))
+end
+
+Then("the product {string} with destination {string} contains resource {string}") do |product, destination, resource|
+  resource_path = Xcode.find_resource(
+    product: product,
+    destination: destination,
+    resource: resource,
+    derived_data_path: @derived_data_path
+  )
+
+  assert(resource_path)
+end
+
+Then("the product {string} with destination {string} does not contain resource {string}") do |product, destination, resource|
+  resource_path = Xcode.find_resource(
+    product: product,
+    destination: destination,
+    resource: resource,
+    derived_data_path: @derived_data_path
+  )
+  refute(resource_path)
 end